Uploading images will always work as expected now.
