+++ src/components/InfoForm.jsx
@@ -1,99 +1,27 @@
 import React from "react";
 import StepFormComponent from "./StepFormComponent";
-import { Upload, Icon, Modal, Form, Input } from "antd";
+import { Modal, Form, Input } from "antd";
+import CustomUpload from "./CustomUpload";
 import TextArea from "antd/lib/input/TextArea";
 import { withFirebase } from "./Firebase";
 
 const initial = "";
 
-function errorMessage(title, content) {
-  Modal.error({ title, content, centered: true });
-}
-
 class InfoForm extends StepFormComponent {
   state = {
     previewVisible: false,
-    previewImage: "",
-    fileList: this.props.formData.fileList ? this.props.formData.fileList : []
+    previewImage: ""
   };
 
   handleCancel = () => this.setState({ previewVisible: false });
 
   handlePreview = file => {
-    console.log("file", file);
+    // TODO: Remove thumburl and preview from URL
+    //  console.log("file", file);
     this.setState({
       previewImage: file.url || file.thumbUrl,
       previewVisible: true
     });
-  };
-
-  handleChange = ({ fileList }) => {
-    console.log("filelist", fileList);
-    this.props.formData.fileList = [...fileList];
-    this.setState({ fileList: this.props.formData.fileList });
-  };
-
-  handleRemove = file => {
-    if (file.originFileObj && file.originFileObj.status == "error") return true;
-    const filename = file.uid + "-" + file.name;
-    this.props.formData.images = this.props.formData.images.filter(
-      image => image !== filename
-    );
-    this.props.firebase
-      .imageUploads()
-      .child(filename)
-      .delete()
-      .then(function() {
-        return true;
-      })
-      .catch(function(error) {
-        errorMessage("Error Removing image:", error.message);
-        return false;
-      });
-  };
-
-  firebaseUpload = file => {
-    if (file.size > 2000000) {
-      // Image > 2mb.
-      file.status = "error";
-      errorMessage(
-        "Image upload error:",
-        "Cannot upload an image over 2mb in size"
-      );
-      return false;
-    }
-    const filename = file.uid + "-" + file.name;
-    this.props.formData.images.push(filename);
-    let uploadTask = this.props.firebase
-      .imageUploads()
-      .child(filename)
-      .put(file);
-
-    uploadTask.on(
-      "state_changed",
-      snapshot => {
-        // Observe state change events such as progress, pause, and resume
-        // Get task progress, including the number of bytes uploaded and the total number of bytes to be uploaded
-        var progress = (snapshot.bytesTransferred / snapshot.totalBytes) * 100;
-        console.log("Upload is " + progress + "% done");
-        file.status = "uploading";
-      },
-      error => {
-        // Handle unsuccessful uploads
-        file.status = "error";
-        file.response = error;
-      },
-      () => {
-        // Handle successful uploads on complete
-        // For instance, get the download URL: https://firebasestorage.googleapis.com/...
-        uploadTask.snapshot.ref.getDownloadURL().then(downloadURL => {
-          console.log("File available at", downloadURL);
-          file.url = downloadURL;
-          file.status = "done";
-        });
-      }
-    );
-    return false;
   };
 
   checkPhoneNumber = (rule, value, callback) => {
@@ -111,7 +39,7 @@ class InfoForm extends StepFormComponent {
   };
 
   checkImages = (rule, value, callback) => {
-    console.log(this.props.formData.images);
+    // console.log(this.props.formData.images);
     if (this.props.formData.images.length < 1) {
       callback("Please upload at least one image.");
     } else {
@@ -120,13 +48,7 @@ class InfoForm extends StepFormComponent {
   };
 
   render() {
-    const { previewVisible, previewImage, fileList } = this.state;
-    const uploadButton = (
-      <div>
-        <Icon type='plus' />
-        <div className='ant-upload-text'>Upload</div>
-      </div>
-    );
+    const { previewVisible, previewImage } = this.state;
 
     const {
       getFieldDecorator,
@@ -198,22 +120,9 @@ class InfoForm extends StepFormComponent {
           required={true}
         >
           <div className='clearfix'>
-            {getFieldDecorator("image", {
+            {getFieldDecorator("images", {
               rules: [{ validator: this.checkImages }]
-            })(
-              <Upload
-                accept='image/*'
-                action=''
-                listType='picture-card'
-                fileList={fileList}
-                onPreview={this.handlePreview}
-                onChange={this.handleChange}
-                beforeUpload={this.firebaseUpload}
-                onRemove={this.handleRemove}
-              >
-                {fileList.length >= 3 ? null : uploadButton}
-              </Upload>
-            )}
+            })(<CustomUpload onPreview={this.handlePreview} {...this.props} />)}
             <Modal
               visible={previewVisible}
               footer={null}
